Move localization interface and default to i18n file
